fix(tests): move VM scanning unit tests to vmhelpers
Move the cluster preflight and VM scan config helpers out of the
e2e-tagged `tests` package so their unit tests run only under the GH
Actions unit-test target. Update the VM scanning suite to use the new
vmhelpers APIs and remove the e2e-tagged unit test files so the VM
e2e job no longer executes those tests.

User request: "Move VM Unit Tests to vmhelpers

Implement the plan as specified, it is attached for your reference.
Do NOT edit the plan file itself.

To-do's from the plan have already been created. Do not create them
again. Mark them as in_progress as you work, starting with the first
one. Don't stop until you have completed all the to-dos."

AI-Assisted: cursor, moved File 1 and File 2 helpers/tests; left File 3 for manual review
