+++ tests/vm_scanning_cluster_preflight_test.go
@@ -9,7 +9,7 @@ import (
 	"testing"
 	"time"
 
-	coreV1 "k8s.io/api/core/v1"
+	"github.com/stackrox/rox/tests/vmhelpers"
 	metaV1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 	"k8s.io/apimachinery/pkg/runtime/schema"
@@ -21,125 +21,14 @@ import (
 var kubeVirtInstallNamespaces = []string{"openshift-cnv", "kubevirt", "openshift-kubevirt"}
 
 const (
-	kvmCapacityResourceName             = coreV1.ResourceName("devices.kubevirt.io/kvm")
-	workerNodeLabel                     = "node-role.kubernetes.io/worker"
-	workerNodeScope                     = "worker-labeled nodes"
-	kvmAllSchedulableNodeScope          = "all schedulable nodes"
-	kvmFallbackDiagnostic               = "No worker-labeled nodes found; checking all schedulable nodes for KVM capacity"
 	vsockFeatureGateWaitTimeout         = 3 * time.Minute
 	vsockVirtHandlerEvidenceWaitTimeout = 3 * time.Minute
 	vsockPreflightPollInterval          = 5 * time.Second
 )
 
-type kvmPreflightNode struct {
-	Name          string
-	Unschedulable bool
-	KVMCapacity   string
-	Eligible      bool
-}
-
-type clusterKVMPreflightResult struct {
-	Scope                           string
-	UsedAllSchedulableNodesFallback bool
-	CheckedNodes                    []kvmPreflightNode
-}
-
-func (r clusterKVMPreflightResult) IsUsable() bool {
-	for _, node := range r.CheckedNodes {
-		if node.Eligible {
-			return true
-		}
-	}
-	return false
-}
-
-func (r clusterKVMPreflightResult) Diagnostic() string {
-	parts := make([]string, 0, 2)
-	if r.UsedAllSchedulableNodesFallback {
-		parts = append(parts, kvmFallbackDiagnostic)
-	}
-
-	if r.IsUsable() {
-		eligibleNodes := make([]string, 0, len(r.CheckedNodes))
-		for _, node := range r.CheckedNodes {
-			if node.Eligible {
-				eligibleNodes = append(eligibleNodes, fmt.Sprintf("%s(kvm=%s)", node.Name, node.KVMCapacity))
-			}
-		}
-		parts = append(parts, fmt.Sprintf("KVM-capable %s: %s", r.Scope, strings.Join(eligibleNodes, " ")))
-		return strings.Join(parts, "\n")
-	}
-
-	checkedNodes := make([]string, 0, len(r.CheckedNodes))
-	for _, node := range r.CheckedNodes {
-		checkedNodes = append(checkedNodes, fmt.Sprintf("%s(unschedulable=%t kvm=%s)", node.Name, node.Unschedulable, node.KVMCapacity))
-	}
-	parts = append(parts, fmt.Sprintf("No %s with devices.kubevirt.io/kvm > 0. Checked: %s", r.Scope, strings.Join(checkedNodes, " ")))
-	return strings.Join(parts, "\n")
-}
-
-func isWorkerNode(node coreV1.Node) bool {
-	_, ok := node.Labels[workerNodeLabel]
-	return ok
-}
-
-func nodeHasPositiveKVMCapacity(node coreV1.Node) bool {
-	quantity, ok := node.Status.Capacity[kvmCapacityResourceName]
-	return ok && quantity.Sign() > 0
-}
-
-func nodeKVMCapacityString(node coreV1.Node) string {
-	quantity, ok := node.Status.Capacity[kvmCapacityResourceName]
-	if !ok {
-		return "<unset>"
-	}
-	return quantity.String()
-}
-
-func inspectClusterKVMReadiness(ctx context.Context, k8s kubernetes.Interface) (clusterKVMPreflightResult, error) {
-	nodeList, err := k8s.CoreV1().Nodes().List(ctx, metaV1.ListOptions{})
-	if err != nil {
-		return clusterKVMPreflightResult{}, err
-	}
-
-	result := clusterKVMPreflightResult{
-		Scope:        workerNodeScope,
-		CheckedNodes: make([]kvmPreflightNode, 0, len(nodeList.Items)),
-	}
-
-	for _, node := range nodeList.Items {
-		if isWorkerNode(node) {
-			result.CheckedNodes = append(result.CheckedNodes, kvmPreflightNode{
-				Name:          node.Name,
-				Unschedulable: node.Spec.Unschedulable,
-				KVMCapacity:   nodeKVMCapacityString(node),
-				Eligible:      !node.Spec.Unschedulable && nodeHasPositiveKVMCapacity(node),
-			})
-		}
-	}
-
-	if len(result.CheckedNodes) == 0 {
-		result.Scope = kvmAllSchedulableNodeScope
-		result.UsedAllSchedulableNodesFallback = true
-		for _, node := range nodeList.Items {
-			if node.Spec.Unschedulable {
-				continue
-			}
-			result.CheckedNodes = append(result.CheckedNodes, kvmPreflightNode{
-				Name:          node.Name,
-				Unschedulable: node.Spec.Unschedulable,
-				KVMCapacity:   nodeKVMCapacityString(node),
-				Eligible:      !node.Spec.Unschedulable && nodeHasPositiveKVMCapacity(node),
-			})
-		}
-	}
-
-	return result, nil
-}
-
 func mustVerifyClusterKVMReady(t testing.TB, ctx context.Context, k8s kubernetes.Interface) {
 	t.Helper()
-	result, err := inspectClusterKVMReadiness(ctx, k8s)
+	result, err := vmhelpers.InspectClusterKVMReadiness(ctx, k8s)
 	if err != nil {
 		t.Fatalf("KVM preflight: could not inspect cluster nodes: %v", err)
 	}
@@ -155,52 +44,23 @@ var kubevirtCRGVR = schema.GroupVersionResource{
 	Resource: "kubevirts",
 }
 
-type kubeVirtVSOCKRef struct {
-	Namespace string
-	Name      string
-}
-
-func verifyClusterVSOCKReadyPhases(
-	ctx context.Context,
-	featureGateTimeout time.Duration,
-	virtHandlerTimeout time.Duration,
-	waitForFeatureGate func(context.Context) (kubeVirtVSOCKRef, error),
-	waitForVirtHandler func(context.Context) (string, error),
-) (kubeVirtVSOCKRef, string, error) {
-	phaseOneCtx, cancelPhaseOne := context.WithTimeout(ctx, featureGateTimeout)
-	ref, err := waitForFeatureGate(phaseOneCtx)
-	cancelPhaseOne()
-	if err != nil {
-		return kubeVirtVSOCKRef{}, "", err
-	}
-
-	phaseTwoCtx, cancelPhaseTwo := context.WithTimeout(ctx, virtHandlerTimeout)
-	lastDiag, err := waitForVirtHandler(phaseTwoCtx)
-	cancelPhaseTwo()
-	if err != nil {
-		return ref, lastDiag, err
-	}
-
-	return ref, lastDiag, nil
-}
-
-func waitForKubeVirtWithVSOCKFeatureGate(ctx context.Context, t testing.TB, dyn dynamic.Interface) (kubeVirtVSOCKRef, error) {
+func waitForKubeVirtWithVSOCKFeatureGate(ctx context.Context, t testing.TB, dyn dynamic.Interface) (vmhelpers.KubeVirtVSOCKRef, error) {
 	t.Helper()
-	var ref kubeVirtVSOCKRef
+	var ref vmhelpers.KubeVirtVSOCKRef
 	err := wait.PollUntilContextCancel(ctx, vsockPreflightPollInterval, true, func(ctx context.Context) (bool, error) {
 		ns, kvObj, _, err := findKubeVirtWithVSOCKFeatureGate(ctx, dyn)
 		if err != nil {
 			t.Logf("VSOCK preflight: KubeVirt CR with VSOCK feature gate not found yet: %v", err)
 			return false, nil
 		}
-		ref = kubeVirtVSOCKRef{
+		ref = vmhelpers.KubeVirtVSOCKRef{
 			Namespace: ns,
 			Name:      kvObj.GetName(),
 		}
 		return true, nil
 	})
 	if err != nil {
-		return kubeVirtVSOCKRef{}, err
+		return vmhelpers.KubeVirtVSOCKRef{}, err
 	}
 	return ref, nil
 }
@@ -209,7 +69,7 @@ func waitForVirtHandlerVsockEvidence(ctx context.Context, t testing.TB, k8s kube
 	t.Helper()
 	var lastDiag string
 	err := wait.PollUntilContextCancel(ctx, vsockPreflightPollInterval, true, func(ctx context.Context) (bool, error) {
-		ok, diag := virtHandlerHostVsockVolumesLookUsable(ctx, t, k8s)
+		ok, diag := vmhelpers.VirtHandlerHostVsockVolumesLookUsable(ctx, t, k8s, kubeVirtInstallNamespaces...)
 		lastDiag = diag
 		return ok, nil
 	})
@@ -220,11 +80,11 @@ func waitForVirtHandlerVsockEvidence(ctx context.Context, t testing.TB, k8s kube
 // It fails with explicit diagnostics suitable for CI triage when either check cannot be satisfied.
 func mustVerifyClusterVSOCKReady(t *testing.T, ctx context.Context, k8s kubernetes.Interface, dyn dynamic.Interface) {
 	t.Helper()
-	ref, lastDiag, err := verifyClusterVSOCKReadyPhases(
+	ref, lastDiag, err := vmhelpers.VerifyClusterVSOCKReadyPhases(
 		ctx,
 		vsockFeatureGateWaitTimeout,
 		vsockVirtHandlerEvidenceWaitTimeout,
-		func(phaseCtx context.Context) (kubeVirtVSOCKRef, error) {
+		func(phaseCtx context.Context) (vmhelpers.KubeVirtVSOCKRef, error) {
 			return waitForKubeVirtWithVSOCKFeatureGate(phaseCtx, t, dyn)
 		},
 		func(phaseCtx context.Context) (string, error) {
@@ -232,7 +92,7 @@ func mustVerifyClusterVSOCKReady(t *testing.T, ctx context.Context, k8s kubernet
 		},
 	)
 	if err != nil {
-		if ref == (kubeVirtVSOCKRef{}) {
+		if ref == (vmhelpers.KubeVirtVSOCKRef{}) {
 			t.Fatalf("VSOCK preflight: no KubeVirt CR with VSOCK feature gate found after %s.\n"+
 				"Tried namespaces %s.\n"+
 				"Remediation: install/configure KubeVirt/CNV and add \"VSOCK\" to spec.configuration.developerConfiguration.featureGates on a KubeVirt CR (see KubeVirt docs for vsock).",
@@ -299,89 +159,4 @@ func findKubeVirtWithVSOCKFeatureGate(ctx context.Context, dyn dynamic.Interface
 		return "", unstructured.Unstructured{}, nil, fmt.Errorf("last list error: %w\n%s", lastListErr, summary)
 	}
 	return "", unstructured.Unstructured{}, nil, fmt.Errorf("%s", summary)
-}
-
-func virtHandlerHostVsockVolumesLookUsable(ctx context.Context, t testing.TB, k8s kubernetes.Interface) (bool, string) {
-	t.Helper()
-	var diag strings.Builder
-	selectors := []string{
-		"kubevirt.io=virt-handler",
-		"app.kubernetes.io/component=virt-handler",
-	}
-	for _, ns := range kubeVirtInstallNamespaces {
-		var pods *coreV1.PodList
-		foundPods := false
-		for _, sel := range selectors {
-			list, lerr := k8s.CoreV1().Pods(ns).List(ctx, metaV1.ListOptions{LabelSelector: sel})
-			if lerr != nil {
-				fmt.Fprintf(&diag, "namespace %q: list pods (%q): %v\n", ns, sel, lerr)
-				continue
-			}
-			if len(list.Items) > 0 {
-				pods = list
-				foundPods = true
-				break
-			}
-		}
-		if !foundPods {
-			fmt.Fprintf(&diag, "namespace %q: no virt-handler pods for selectors %v\n", ns, selectors)
-			continue
-		}
-		for i := range pods.Items {
-			pod := &pods.Items[i]
-			phaseReady := pod.Status.Phase == coreV1.PodRunning || pod.Status.Phase == coreV1.PodPending
-			if !phaseReady {
-				fmt.Fprintf(&diag, "namespace %q pod %q: phase=%q\n", ns, pod.Name, pod.Status.Phase)
-			}
-			hasExplicitVsockPath := false
-			hasCNVLibvirtRuntimePath := false
-			for _, vol := range pod.Spec.Volumes {
-				if vol.HostPath == nil {
-					continue
-				}
-				p := strings.ToLower(vol.HostPath.Path)
-				if strings.Contains(p, "vsock") {
-					hasExplicitVsockPath = true
-					break
-				}
-				if strings.Contains(p, "kubevirt-libvirt-runtimes") {
-					hasCNVLibvirtRuntimePath = true
-				}
-			}
-			if phaseReady {
-				if hasExplicitVsockPath {
-					return true, diag.String()
-				}
-				// OCP Virtualization commonly exposes vsock plumbing through libvirt runtime mounts
-				// without an explicit "/dev/vhost-vsock" hostPath in the pod spec.
-				if hasCNVLibvirtRuntimePath {
-					fmt.Fprintf(&diag, "namespace %q pod %q: accepting CNV libvirt runtime hostPath as vsock evidence\n", ns, pod.Name)
-					return true, diag.String()
-				}
-			}
-			fmt.Fprintf(&diag, "namespace %q pod %q: hostPath volumes (no vsock-like path): %s\n",
-				ns, pod.Name, summarizePodHostPathVolumes(pod))
-		}
-	}
-	return false, diag.String()
-}
-
-func summarizePodHostPathVolumes(pod *coreV1.Pod) string {
-	var b strings.Builder
-	n := 0
-	for _, vol := range pod.Spec.Volumes {
-		if vol.HostPath == nil {
-			continue
-		}
-		if n > 0 {
-			b.WriteString("; ")
-		}
-		// Fprintf to a Builder avoids the intermediate string allocation of Sprintf+WriteString.
-		fmt.Fprintf(&b, "%s->%q", vol.Name, vol.HostPath.Path)
-		n++
-	}
-	if n == 0 {
-		return "<none>"
-	}
-	return b.String()
 }
