+++ tests/vmhelpers/vm_scan_config.go
@@ -1,6 +1,4 @@
-//go:build test_e2e_vm
-
-package tests
+package vmhelpers
 
 import (
 	"crypto/ed25519"
@@ -39,31 +37,33 @@ var (
 	repo2CPEURL = env.RegisterSetting("ROXAGENT_REPO2CPE_URL", env.WithDefault(defaultRepo2CPEURL))
 )
 
-// vmSpec describes a VM to provision: container-disk image and guest SSH user.
-type vmSpec struct {
+// VMSpec describes a VM to provision: container-disk image and guest SSH user.
+type VMSpec struct {
 	Name      string
 	Image     string
 	GuestUser string
 }
 
-type vmScanConfig struct {
-	Images              []string // container-disk images (from VM_IMAGES, comma-separated)
-	GuestUsers          []string // per-image SSH users (from VM_USERS, comma-separated; shorter lists are padded with defaultGuestUser)
+// VMScanConfig contains the VM scanning test configuration derived from the environment.
+type VMScanConfig struct {
+	Images              []string
+	GuestUsers          []string
 	VirtctlPath         string
 	RoxagentBinaryPath  string
 	Repo2CPEURL         string
-	SSHPrivateKey       string // PEM-encoded private key content (not a file path)
-	SSHPublicKey        string // OpenSSH authorized_keys line (not a file path)
+	SSHPrivateKey       string
+	SSHPublicKey        string
 	NamespacePrefix     string
 	ScanTimeout         time.Duration
 	ScanPollInterval    time.Duration
 	DeleteTimeout       time.Duration
 	SkipCleanup         bool
-	ImagePullSecretPath string // Path to docker config JSON for private registries
+	ImagePullSecretPath string
 }
 
-func loadVMScanConfig() (*vmScanConfig, error) {
-	cfg := &vmScanConfig{}
+// LoadVMScanConfig reads the VM scanning configuration from the environment.
+func LoadVMScanConfig() (*VMScanConfig, error) {
+	cfg := &VMScanConfig{}
 
 	var err error
 	imagesRaw := strings.TrimSpace(os.Getenv("VM_IMAGES"))
@@ -86,7 +86,7 @@ func loadVMScanConfig() (*vmScanConfig, error) {
 			cfg.GuestUsers = append(cfg.GuestUsers, strings.TrimSpace(u))
 		}
 	}
-	if cfg.VirtctlPath, err = discoverVirtctlPath(); err != nil {
+	if cfg.VirtctlPath, err = DiscoverVirtctlPath(); err != nil {
 		return nil, err
 	}
 	if cfg.RoxagentBinaryPath, err = discoverRoxagentBinaryPath(); err != nil {
@@ -99,7 +99,7 @@ func loadVMScanConfig() (*vmScanConfig, error) {
 	cfg.SSHPublicKey = strings.TrimSpace(os.Getenv("VM_SSH_PUBLIC_KEY"))
 	switch {
 	case strings.TrimSpace(cfg.SSHPrivateKey) == "" && cfg.SSHPublicKey == "":
-		priv, pub, genErr := generateEphemeralSSHKeypair()
+		priv, pub, genErr := GenerateEphemeralSSHKeypair()
 		if genErr != nil {
 			return nil, fmt.Errorf("VM_SSH_PRIVATE_KEY/VM_SSH_PUBLIC_KEY not set and ephemeral key generation failed: %w", genErr)
 		}
@@ -125,16 +125,15 @@ func loadVMScanConfig() (*vmScanConfig, error) {
 	return cfg, nil
 }
 
-// vmSpecs builds the VM specification list from the parsed images and guest
-// users. VM names are generated as vm-0, vm-1, etc.
-func (c *vmScanConfig) vmSpecs() []vmSpec {
-	specs := make([]vmSpec, len(c.Images))
+// VMSpecs builds the VM specification list from the parsed images and guest users.
+func (c *VMScanConfig) VMSpecs() []VMSpec {
+	specs := make([]VMSpec, len(c.Images))
 	for i, img := range c.Images {
 		user := defaultGuestUser
 		if i < len(c.GuestUsers) && c.GuestUsers[i] != "" {
 			user = c.GuestUsers[i]
 		}
-		specs[i] = vmSpec{
+		specs[i] = VMSpec{
 			Name:      fmt.Sprintf("vm-%d", i),
 			Image:     img,
 			GuestUser: user,
@@ -143,8 +142,8 @@ func (c *vmScanConfig) vmSpecs() []vmSpec {
 	return specs
 }
 
-// discoverVirtctlPath returns the VIRTCTL_PATH env var if set, otherwise searches $PATH.
-func discoverVirtctlPath() (string, error) {
+// DiscoverVirtctlPath returns the VIRTCTL_PATH env var if set, otherwise searches $PATH.
+func DiscoverVirtctlPath() (string, error) {
 	if v := strings.TrimSpace(os.Getenv("VIRTCTL_PATH")); v != "" {
 		info, err := os.Stat(v)
 		if err != nil {
@@ -162,8 +161,6 @@ func discoverVirtctlPath() (string, error) {
 	return p, nil
 }
 
-// discoverRoxagentBinaryPath returns the ROXAGENT_BINARY_PATH env var if set,
-// otherwise probes the standard build output path relative to the repository root.
 func discoverRoxagentBinaryPath() (string, error) {
 	if v := strings.TrimSpace(os.Getenv("ROXAGENT_BINARY_PATH")); v != "" {
 		return v, nil
@@ -176,7 +173,6 @@ func discoverRoxagentBinaryPath() (string, error) {
 	return "", fmt.Errorf("ROXAGENT_BINARY_PATH not set and %s does not exist; run 'make roxagent_linux-amd64'", candidate)
 }
 
-// repoRoot returns the repository root by walking up from this source file.
 func repoRoot() string {
 	_, file, _, ok := runtime.Caller(0)
 	if !ok {
@@ -185,9 +181,8 @@ func repoRoot() string {
 	return filepath.Clean(filepath.Join(filepath.Dir(file), ".."))
 }
 
-// generateEphemeralSSHKeypair creates a one-time ed25519 keypair and returns
-// the PEM-encoded private key and the OpenSSH authorized_keys public key line.
-func generateEphemeralSSHKeypair() (privateKeyPEM string, publicKeyAuthorized string, err error) {
+// GenerateEphemeralSSHKeypair creates a one-time ed25519 keypair and returns the key material.
+func GenerateEphemeralSSHKeypair() (privateKeyPEM string, publicKeyAuthorized string, err error) {
 	pub, priv, err := ed25519.GenerateKey(rand.Reader)
 	if err != nil {
 		return "", "", fmt.Errorf("generate ed25519 key: %w", err)
@@ -207,7 +202,8 @@ func generateEphemeralSSHKeypair() (privateKeyPEM string, publicKeyAuthorized st
 	return string(pemData), authorizedKey, nil
 }
 
-func mustFindExecutable(t *testing.T, name string) string {
+// MustFindExecutable resolves an executable on PATH for tests that need a known-good binary.
+func MustFindExecutable(t *testing.T, name string) string {
 	t.Helper()
 
 	path, err := exec.LookPath(name)
